requirements.txt
pwntools
colorama
datetime
